SIB-15 Fix naming: renamed project
