+++ src/main/java/software/sigma/bu003/internship/coparts/CopartsApplication.java
@@ -1,13 +1,13 @@
-package software.sigma.bu003.internship.vedmid_andrii;
+package software.sigma.bu003.internship.coparts;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class VedmidAndriiApplication {
+public class CopartsApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(VedmidAndriiApplication.class, args);
+		SpringApplication.run(CopartsApplication.class, args);
 	}
 
 }
