+++ src/main/java/software/sigma/bu003/internship/coparts/repository/PartRepository.java
@@ -1,8 +1,8 @@
-package software.sigma.bu003.internship.vedmid_andrii.repository;
+package software.sigma.bu003.internship.coparts.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-import software.sigma.bu003.internship.vedmid_andrii.entity.Part;
+import software.sigma.bu003.internship.coparts.entity.Part;
 
 @Repository
 public interface PartRepository extends MongoRepository<Part, String> {
