+++ src/test/java/software/sigma/bu003/internship/coparts/controller/PartControllerTest.java
@@ -1,4 +1,4 @@
-package software.sigma.bu003.internship.vedmid_andrii.controller;
+package software.sigma.bu003.internship.coparts.controller;
 
 import com.mongodb.MongoWriteException;
 import org.junit.jupiter.api.Test;
@@ -8,9 +8,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-import software.sigma.bu003.internship.vedmid_andrii.entity.Part;
-import software.sigma.bu003.internship.vedmid_andrii.service.PartService;
-import software.sigma.bu003.internship.vedmid_andrii.service.exception.PartNotFoundException;
+import software.sigma.bu003.internship.coparts.entity.Part;
+import software.sigma.bu003.internship.coparts.service.PartService;
+import software.sigma.bu003.internship.coparts.service.exception.PartNotFoundException;
 
 import javax.validation.ConstraintViolationException;
 import java.util.List;
