+++ src/test/java/software/sigma/bu003/internship/coparts/service/PartServiceTest.java
@@ -1,9 +1,9 @@
-package software.sigma.bu003.internship.vedmid_andrii.service;
+package software.sigma.bu003.internship.coparts.service;
 
 import org.junit.jupiter.api.Test;
-import software.sigma.bu003.internship.vedmid_andrii.entity.Part;
-import software.sigma.bu003.internship.vedmid_andrii.service.exception.PartNotFoundException;
-import software.sigma.bu003.internship.vedmid_andrii.repository.PartRepository;
+import software.sigma.bu003.internship.coparts.entity.Part;
+import software.sigma.bu003.internship.coparts.service.exception.PartNotFoundException;
+import software.sigma.bu003.internship.coparts.repository.PartRepository;
 
 import java.util.List;
 import java.util.Optional;
