--- src/test/java/software/sigma/bu003/internship/vedmid_andrii/VedmidAndriiApplicationTests.java
@@ -1,13 +1,0 @@
-package software.sigma.bu003.internship.vedmid_andrii;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class VedmidAndriiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
